Refactor lookup functions
